feat: project organization
